--- a/.gitignore
+++ b/.gitignore
@@ -1,4 +1,5 @@
 target
 Cargo.lock
 *.swp
+*.swo
 .cargo