--- a/src/keys.rs
+++ b/src/keys.rs
@@ -10,7 +10,14 @@ use NibbleVec;
 /// encountering a conflict. Be careful!
 pub trait TrieKey: PartialEq + Eq {
     /// Encode a value as a vector of bytes.
-    fn encode(&self) -> Vec<u8>;
+    fn encode_bytes(&self) -> Vec<u8> {
+        panic!("implement this method or TrieKey::encode");
+    }
+
+    /// Encode a value as a NibbleVec.
+    fn encode(&self) -> NibbleVec {
+        NibbleVec::from_byte_vec(self.encode_bytes())
+    }
 }
 
 /// Key comparison result.
@@ -27,16 +34,17 @@ pub enum KeyMatch {
 }
 
 /// Compare two Trie keys.
-pub fn match_keys(first: &NibbleVec, second: &NibbleVec) -> KeyMatch {
-    let min_length = ::std::cmp::min(first.len(), second.len());
+pub fn match_keys(start_idx: usize, first: &NibbleVec, second: &NibbleVec) -> KeyMatch {
+    let first_len = first.len() - start_idx;
+    let min_length = ::std::cmp::min(first_len, second.len());
 
     for i in 0..min_length {
-        if first.get(i) != second.get(i) {
+        if first.get(start_idx + i) != second.get(i) {
             return KeyMatch::Partial(i);
         }
     }
 
-    match (first.len(), second.len()) {
+    match (first_len, second.len()) {
         (x, y) if x < y => KeyMatch::FirstPrefix,
         (x, y) if x == y => KeyMatch::Full,
         _ => KeyMatch::SecondPrefix
@@ -51,9 +59,8 @@ pub fn check_keys<K>(key1: &K, key2: &K) where K: TrieKey {
 }
 
 /// --- TrieKey Implementations for standard types --- ///
-
 impl<T> TrieKey for T where T: Into<Vec<u8>> + Clone + Eq + PartialEq {
-    fn encode(&self) -> Vec<u8> {
+    fn encode_bytes(&self) -> Vec<u8> {
         self.clone().into()
     }
 }