This is way faster, and we get shiny new APIs.