@michaelsproul 2016-10-08
This is disturbingly much faster with `rustc 1.14.0-nightly (9c31d76e9 2016-10-03)`.

Old:

```
test benchmark::trie_get           ... bench: 121,716,126 ns/iter (+/- 8,098,068)
test benchmark::trie_insert        ... bench: 143,402,909 ns/iter (+/- 24,101,428)
test benchmark::trie_insert_remove ... bench: 249,433,134 ns/iter (+/- 10,662,332)
```

New:

```
test benchmark::trie_get           ... bench:  30,220,374 ns/iter (+/- 3,973,437)
test benchmark::trie_insert        ... bench:  52,295,969 ns/iter (+/- 4,604,410)
test benchmark::trie_insert_remove ... bench:  81,922,786 ns/iter (+/- 9,844,586)
```