@andrewcsmith 2016-10-08
This is very cool. I like what I see so far. One major roadblock is that there is no common trait that both `Trie` and `SubTrie` implement. It would be great to abstract many of the operations so that any a given function can take either struct.

Second, I pushed up a new commit to your new-stuff branch with the changes I had to make in order to get things working with my current library (a voting-experts algorithm library, if you must know). It turned out to be very minimal!

The voting experts part is not the bottleneck of my application, but I'll use it to do some text-segmentation on a fully concatenated version of Moby Dick and let you know if I see a performance improvement there.