@michaelsproul 2016-10-08
Great!

Good idea regarding the trait! I think we should also keep `TrieNode` private and just add wrappers for the methods that are currently missing (the iterators, maybe subtries of subtries, but I don't really want to tackle that yet).