@michaelsproul 2016-10-08
Those benchmarks from before were using the OS X system allocator, without it, the old code performs a bit better:

## Old

```
test benchmark::trie_get           ... bench:  82,447,514 ns/iter (+/- 5,546,543)
test benchmark::trie_insert        ... bench:  71,935,831 ns/iter (+/- 4,707,764)
test benchmark::trie_insert_remove ... bench: 130,737,040 ns/iter (+/- 34,883,760)
```

## New

```
test benchmark::trie_get           ... bench:  31,841,361 ns/iter (+/- 9,877,950)
test benchmark::trie_insert        ... bench:  34,958,108 ns/iter (+/- 6,714,198)
test benchmark::trie_insert_remove ... bench:  62,985,124 ns/iter (+/- 10,797,948)
```

Ratios are: 2.59 times faster for get, 2.06 for insert, 2.07 for remove.