@michaelsproul 2016-10-08
I've added a `TrieCommon` trait that unifies Trie/SubTrie/SubTrieMut. Due to some funky stuff with mutable borrows, there are *two* implementations related to `SubTrieMut`, one for `&'b SubTrieMut<'a, K, V>` from which only values/subtries/etc with lifetime `'b` can be extracted, and one for `SubTrieMut<'a, K, V>` from which values with lifetime `'a` (the original Trie lifetime) can be extracted.