@michaelsproul 2016-10-08
The core reason for the difficulty with mutable subtries is that it's impossible to write this function:

```rust
fn borrow<'a: 'b, 'b>(x: &'b &'a mut String) -> &'a String {
    &**x
}
```

The reason for this is that it results in multiple pointers to the underlying string, which violates the property that the `&mut String` is the only reference.