@andrewcsmith 2016-10-09
Cool. I'm looking at it, but it's taking a sec. I think I need to redesign bits of my implementation. Sorry—it's very hard for me to stare at code that's as complex as this subtrie borrowing mechanism without implementing something and seeing if it works.