@michaelsproul 2016-10-10
Yeah no worries, it's a bit insane the subtrie borrowing thing...