@andrewcsmith 2016-10-11
Excellent. It took a little while, but I got everything working without touching the changes you made. I think we're good. At least, I'm satisfied.

I do agree that we need the option of borrowing the key with a `Cow`. I doubt it's going to kill performance at this point, but it would be cool to avoid clones when possible.

I'll try and do the moby dick segmentation pretty soon, to get some benchmarks going with my application.