@andrewcsmith 2016-10-15
Hm, I have tests written for all my stuff so I'm not seeing any failures. I don't think I'm working around anything either, but I'm also not matching keys on the `SubTrie`s. (I'm only using the `Children` iterator method.)

[These two methods](https://github.com/andrewcsmith/voting_experts/blob/master/src/trie.rs#L70) (`boundry_entropy` and `boundary_entropy_recur`) are the most difficult parts for me, because of the borrowing.

Incorporating your changes now, and I'll take a closer look when I hack at my own code too.