@andrewcsmith 2016-10-15
Now I see your point with the prefix. The changes make sense to me.

I'm going to see what I can do about getting the memory usage down by using `Cow` where possible. It could create some borrow checker hell of course.

Also pushed an example of String frequency evaluation.