@michaelsproul 2016-10-16
Ok, gonna merge this to master so we can work on it more easily.